feat: make FFT and Multiexp more flexible
This library was built for being used with BLS12-381. This commit
makes the FFT and Multiexp implementations a bit more flexible.
It's now possible to use it also for other curves/fields. You
can even define several FFT and/or Multiexp kernels within the
same source file.

To create a kernel to be used with BLS12-381 would look like this:

    use blstrs::{Fp, Fp2, G1Affine, G2Affine, Scalar};
    use ec_gpu_gen::{Limb32, SourceBuilder};

    let source = SourceBuilder::new()
        .add_fft::<Scalar>()
        .add_multiexp::<G1Affine, Fp>()
        .add_multiexp::<G2Affine, Fp2>()
        .build_32_bit_limbs();

BREAKING CHANGE: Kernel source generation has a different API now

FFT and Multiexp are no longer BLS12-381 specific. This also means
that the `multiexp` and `fft` features were removed. You now need
to define a kernel within your library.